python: Split colteconf script into network and billing components
This commit starts the process of splitting the colteconf script into
a part that manages the configuration of the core network, and a part
that manages the billing system configuration. Both parts reference
the same unified colte config file. This commit lays the foundation
for future commits actually splitting up the installed packages.
